MINOR: cb: simplify the descent API a little bit
Instead of the descent function updating node->l and node->r, it now
returns the side the leaf should occupy on the new node (which becomes
the parent) so that the insert function can proceed by itself. That's
less arguments to pass and it becomes more natural to use. The code is
also 8 bytes smaller.