MINOR: cb: fix delete segfault
The problem with 1 2 1 is that the first 1 has no node and does not match
any of the conditions expected and we end up dereferencing a null nparent.
By initializing it to the root, we seem to address this but it would be
important to double-check if that's sufficient or not, because:
  - nowhere we explicitly deprived 2 from its node, yet it worked fine
  - nowhere we fixed 2's leaves to have the two pointing to the leaf

So maybe it's still not completely fixed. And indeed:

    1 2 1 3 2 4 1 2 1 2 3 1 4 1 1 2 2 2

ends up with only 2 nodes.