EXP: now lookup is significantly faster, but...
The problem is that as soon as we use root = &p->node.b[brside] during the
descent, then the lookups go from 5.8s to 6.0s because the compiler reorders
the loop. In fact without the prefetches and the early comparison, it's
even worse (6.4s).

This should likely be experimented on top of memory-based contents so that
we decide afterwards in which order to perform the comparisons.