EXP: add a pick() operation
This one is more convenient than lookup+delete since it doesn't require
two lookups. However it is difficult to implement because the node passed
in argument for the lookup is not the one from the tree, so we can't
compare node pointers to assign the node's parent, we have to use the
key instead. That's cheap for scalars, but will we be able to do that
with strings? Also we don't know how that will work with dups. Maybe we
should clearly detect early matches for unique trees instead of going
down to the leaf after all. An alternate approach could be to implement
a stack of visited nodes and look back for duplicates. A bit complex
though.