TESTS: add speedcbamb.c to perform many insertions and lookups
This is used to insert and lookup data. It uses a u64 PRNG and uses the
same sequence for each loop. This way by looking up less values it's
possible to measure the cost of hits, and by looking up more values it's
possible to measure the cost of misses as well. E.g:

  $ time tests/speedcbamb 1000000 2000000 10

This will insert 1 million values, and look up 2 million ones, the first
1 million of which are hits, and loops 10 times over these lookups.