MINOR: implement the string mode
This is a direct string mode. It works like the memory block mode except
that the data stops at the trailing zero (which is always stored).

Unexpectedly the approach is quite more difficult here because of the
unknown length of the elements: for example we don't know if the
branches' keys are before or after the current one so they may very
well be shorter, which prevents us from comparing from a known restart
point, or even checking for the end of the string by looking at a given
offset. This requires to use strcmp() at various places, restarting from
a known safe offset.

We manage to avoid a few occasional comparisons by marking that the
entry was found in certain cases. It might be possible to improve this
so that any time the entry is found we keep that info to avoid the late
checks, but it doesn't seem to significantly improve the performance.
Another option could be to store te strcmp() result since we need to
know the sign at the end. This could be experimented with.

Regardless, with 20-char keys, this achieves 3M lookups per second among
100k keys.

Maybe a new comparison method could improve the situation as well: instead
of comparing K with L, then K with R, then L with R, it could make sense
to compare K,L,R together and report all needed elements. This would read
K,L,R only once instead of twice. It's just not sure how much could be
saved given that the second read is from the cache.