MINOR: cb: add a walk method definition
The cba_walk_meth enum will be used to indicate how we want to descend
a tree:
  - walk left (first)
  - walk right (last)
  - match the key

This remains relevant because it allows to handle all the special cases
of the multiple return values. For example we'll be able to implement
the pick_first() function and so on by simply reusing the unique descent
function.