MINOR: st: return the last left or right alternate sub trees if desired
The descent function now can fill a caller-passed pointer with the last
right sub-tree or the last left sub-tree every time a direction is
chosen. The goal is to permit the caller to restart a left or right
walk along this path to implement a next/prev/gt/lt etc.