BUG: st: fix the next/prev mechanisms
It didn't work because we need to have the previous node's plen to
compare xlen against, otherwise we'll loop forever on a node that's
also its own leaf, or will take a wrong direction for the first step.

E.g:
        0       # after finding leaf 0, the right branch would be &1.b[1]
      /   \     # and would then always return zero when going down left.
     1     2
   /   \
  0     1

By saving the fork's parent, we save the upper node (0), noting the fork
point at 1 and restart from 0 above, carefully taking the other branch
when the previous branch was mentioned, then going down left all along.

Another approach might be to pass plen in argument to the function, which
could avoid comparisons on the way down. But it would also be needed to
return it.