MINOR: st: experiment by passing plen during next
There seems to be a real difficulty in detecting that a node is a leaf,
particularly when the pointer is equal to one of its own branches, and
to find the last turn point. Indeed, if we fail to see that we're
restarting from the node's own branch, we don't detect it as a leaf
and we can loop over the leftmost branch. The previous method was more
reliable.

The problem is that we need to detect we're on a leaf, which implies to
return the current node's pointer (i.e. root) and not its branch, and that
the walk method must be distinct to explicitly indicate NXT instead of
FST, so that we first take the right and then take only the left branch.

This is needed so that the plen can be calculated and compared to xlen.
We don't even need to return plen, as long as the "next" operation is
returned.

Note that the code changes the way plen is used, and insists on using
xlen instead, which seems to make the code clearer. It doesn't appear
to change its efficiency. It works fine now.