BROKEN: st: experiment by passing plen during next
This sort of works but it skips 90% of the elements. There seems to be
a real difficulty in detecting that a node is a leaf, particularly when
the pointer is equal to one of its own branches, and to find the last
turn point. Indeed, if we fail to see that we're restarting from the
node's own branch, we don't detect it as a leaf and we can loop over
the leftmost branch. The previous method was more reliable.

However the code changes the way plen is used, and insists on using xlen
instead, which seems to make the code clearer. It doesn't appear to
change its efficiency.