FIX: st: OK now it works with plen.
The problem is that we need to detect we're on a leaf, which implies to
return the current node's pointer (i.e. root) and not its branch, and that
the walk method must be distinct to explicitly indicate NEXT instead of
LEFT, so that we first take the right and then take only the left branch.

This is needed so that the plen can be calculated and compared to xlen.
We don't even need to return plen, as long as the "next" operation is
returned.