WIP: st: don't pass node unless inserting.
We used to pass the node all the time, even to have the key. But
the functions used do not employ the node, requiring another cast to
retrieve the key back. Now insted we just pass the key separately.

This will ease the replacement of the casts with plain offsets so
that nodes can be used with any key in any structure.