REORG: create a private include to gather internal macros, types and functions
Since we're heavily relying on a large generic, multi-purpose, inline
function for the descent and that many arguments will need to be passed
to describe how to operate, we'll need more internal types. Let's create
a new private include file for this, and already move cba_walk_meth there.
This file must not be included by user code since it may contain short
names that are subject to clash with user code.