MINOR: descent: pass a key type argument
The key type for now only supports string or none, and is associated
with the key which was renamed key_ptr as it will only be used this
way for pointers. Other key types will come.

The key type indicates what argument to use when a key is provided for
meth==CB_WM_KEY, and how to interpret the contents of the tree during
the descent (how to compare nodes, how to display them in debug mode etc).