MINOR: next/prev: make the next and prev tree walk functions generic
The key is needed since prev/next start by a lookup, so the functions
will have mostly an API compatible with a lookup function, which is
expected given that in practice they're really performing a lookup.
It's likely that in the future the node will be needed in the generic
functions when dups will be needed though.