MINOR: delete: make the delete code generic
It's much cleaner this way, the type-specific function is no different
from a lookup function. It just needs to pass the node and the key.

An extra test was added to the generic function so that it can ignore
the node when it's NULL and only focus on the key. This will permit to
implement the pick() function at the same time. Interestingly, the test
on the non-null ret managed to reduce the code size.