MINOR: insert: return NULL instead of the node on success
At the lower level, in _cbau_descend(), we had an exception for the
insert code to return the node when the key was not found. That's
pointless, the caller knows the node, so better generalize the return
of the NULL to indicate the key is not there, as an invitation to add
it. This reduces the insert code by 0x20 bytes!