MINOR: descent: implement the u32 descent in the generic function
Based on the key type, we can reintegrate the u32 descent there in the
same code structure. The code becomes ugly but the comments are valid
for all types during each phase of the descent, and permit to compare
what is done in which case, so that the logic is preserved.