MEDIUM: u32: now rely on the generic functions
The code is much simpler now and immediately benefits from the utility
operations such as first/next etc. The old code that was commented out
for lookup_ge/le was preserved as for now we don't strictly have the
equivalent, it will serve for illustration purposes later (even though
it comes from the version that only organizes nodes based on their own
pointers).

The dump code was preserved because it works. Later it should be remerged
into the generic code, though the walk method is more complicated there.