MINOR: descent: implement the fixed-size memory block (mb)
The memory block is passed with its length in key_u64 and the pointer
in key_ptr. It was verified that even on 32-bit archs, passing a size
there in a u64 will not cause a cast nor sign extensions.

A few casts were necessary because plen/xlen etc are of type "int". Maybe
switching them to ssize_t could improve the situation at no cost, since
all sizes are castable to size_t anyway. Another approach would consist
in having a size_t klen variable inside the function that's taken from
the argument.