MINOR: delete: always reset the left branch when deleting
The delete() function checks that the left branch is not NULL when
deleting but it still didn't reset it upon operation. Let's do it
now, make sure to return NULL when searching for a deleted node, and
add a cb_intree() function to easily check if a node is still in the
tree.