MINOR: cb: add support for unsigned long keys
Quite often in code we're using unsigned long values without wanting to
be bothered with the key size (e.g. store a size etc). Let's have official
support for unsigned long keys via cbul_*. These are mapped to either 32
or 64 bit calls depending on the size of a long. In order to ease code
usage, they do not remap to the existing 32/64 bit functions so that there
is no porting problems when some types were not linked in for example.