BUG: descent: always update the last xor between integer branches
For the sake of optimizations, in strings and memory blocks, we would
immediately save the common length between branches when it's measured,
in xlen, and the previous one is stored in plen. But for integers we
didn't save it as it was cheap to compute where needed. But that causes
a problem for keyless walks (prev, next etc) because in this case the
previous value (pxor) is never updated. This breaks the "next" walk
across trees arranged like the one below resulting from the insertion
of 64,32,16 in this sequence (N=node, L=leaf):

          root
           |
          N32
         /   \
       N16   L64
      /   \
    L16   L32

In this case, next(16) remains on 16 because we restart from &N32.b[0]
with pxor=~0, and it's never updated when descending to N16, so what
happens is that when reaching L32(=N32) we would continue down to N16
and stop on L16=N16 detected as a leaf. This also explains why the
issue was not detected on string-based operations but only on integers.

The change in this patch makes sure that the XOR between two integer
branches is always saved in xor32/64 and that pxor32/64 is updated
every time we descent one level, like it is done for plen from xlen.
Thanks to this, it's now visible that xor(L32)=80 is higher than
xor(N16)=48 and we can stop.