MINOR: cb: add new CB_WM_* values to support different walk methods
Till now we only supported first/last/next/prev and key to look up an
exact key. However in order to implement lookup_le and lookup_ge we'll
also need to support variants of the key, precisely the LE and GE
variants.

This commit renames CB_WM_KEY to CB_WM_KEQ (for equal) and reserves
_KLE and _KGE for the LE and GE tests.