BUG: lookup: skip root for LE/GE fallback
Otherwise sometimes we can restart from the wrong branch.

But that's still not sufficient:

   ./tests/testcbul -g 40 60 50 51 52

Lookup_ge(52) after 40,60,50,51 retrieves 51 because it goes down the
closest branch (50/51) and the last recorded turn is too late, it
corresponds to split 50/60.branch[0], so it cannot visit the right
side.

In fact it's more complicated. We *do* visit the right side, but the
lower node is the same as the parent node, and since there's no known
xor at this level, we go down the wrong branch against past it instead
of stopping immediately below the root:

            R
            |
            60
           /  \
         40    50
              /  \
             51  60
            /  \
           50  51

Lookup(52) finds nothing and restarts from node 50, which goes down the
left side (51) then full right. There's no way there to go down 60. If
we change the way the right_branch is assigned to point to 50.right,
then we descend along 60,50,51. It's again impossible to perform a
next() on such a node that's taken for the root, since the root only
has a left branch, so the right_branch should not be assigned to node_50
but to node_60.