TESTS: ul: completely rework the stress testing tool
The previous tests were very limited. Let's rework them to use a table
and will-width long ints instead. We can visit random places in the
table to verify if the items there are in the tree or not, to visit
their predecessors and successors, to insert new entries and remove
old ones, etc. We crash upon failure. For now this manages to spot bugs
in the implementation. And it's multi-threaded so that it's faster to
run complex tests.