DEBUG: descent: switch to the new dbg() function
This removes 1/6 of the lines and 1/4 of the size of the whole
cbtree-prv.h file, by getting rid of the hideous macros all over
the descent code. The debug code will likely be a bit less efficient
since it may recalculate lengths or xors that are already known, but
it will no longer depend on where they are set, not will require long
switch/case blocks along the descent. The output is also better
formatted and more readable.