CLEANUP: descent: group the descent by type
There are several chained operations in the descent that are
type-specific, and having them checked by type in small steps is harder
to follow than having larger blocks for a given type.

This patch groups the brside calculation with the check for a leaf
that's based on l->key and r->key as well, and which allows to reduce
the visual scope of certain variables in the descent. The code remains
exactly the same though.