CLEANUP: descent: move the temporary inter-branch xor into their own blocks
This removes some of the function-wide variables and makes it clearer
that their scope is really limited to the check against pxor. This also
moves the pxor assignment to the end of the same block, which has a tiny
side effect in debug mode, which is that the side0 and side1 debug outputs
will contain the new pxor values. The case where p==root doesn't really
change because in this case the pxor would end up being the same anyway.