MINOR: descent: prepare for adding new key-based walk methods
We're now placing CB_WM_KEY at the end so that we can check for a key's
presence using only >= CB_WM_KEY, which will simplify the addition of new
key-based walk modes.

Also let's clarify the conditions at the end of the loop, what is
granted, what is not, etc.