EXP: descent: return the back node differently now
EXP: loop inside the function for next/prev

That's only for u64 for now. The result is slightly smaller but also slower.

The construct is not very convenient but allows to get rid of annoying
args. It also shows what we really need to keep (a pointer to the last
node, and a copy of the pxor, though that's one for each type). It also
requires an early check before entering the loop to make sure we detect
the leaf that depends from the same node, otherwise we can't detect it
later.

Maybe at some point we should instead define a context to be reused
between iterations instead of passing that many pointers. The only
issue is that it will no longer be possible to eliminate useless writes
to such variables we don't need.

EXP: descent: take the initial pxor from the key for NXT/PRV

This will allow some double-step calls to pass the previous value back.
The code didn't change.

It looks like it would be easier done using NXT/PRV than LST/FST:

  - with NXT/PRV, root could be passed as &p->node.b[0], then
    p would be derived from it, then depending on NXT/PRV, root
    would become &p->node.b[0/1] and the check could be done.

  - the return value would only be the node to jump back to

  - the side to be chosen is decided from the operations so we
    don't need to return it

  - we need to read the initial values from the args but that's OK

  - we also need to return xor32/xor64/xlen => that's 3 extra args,
    possibly 2 if we use xor64 for xlen.

EXP: descent: finally just return the return node

Now we only return the return node on KNX/KPR, and pass a pointer to
it as the root pointer for next lookup using NXT/PRV, which will do the
right thing. It will measure the distance between the node's two branches
again, but saves us from having to save the intermediary XOR values. It
still ends up being as fast as the original without requiring access to
the left/right branches.