MINOR: descent: return a single pointer to the back node
After different approaches were tried, it was obvious that while the
key lookup method is obviously the same for regular lookup, next and
prev, it's not the same anymore for lookup_le/lt/ge/gt and we'll need
the descent code to know why we're descending anyway.

So let's take a different approach and implement new lookup methods
"KNX" for "key next" and "KPR" for "key prev". For now they're doing
exactly the same, except that each of them is responsible for deciding
to keep a copy of the last node which turned left or right during the
descent. This way we don't need to pass two pointers anymore, a single
one for the backup node is sufficient, which simplifies the callers a
little bit.

One nice benefit in this is that the descent code continues to return
the looked up node in addition to the restart node. This way the
construct is cleaner and more robust because the function's return
indicates whether or not an entry was found for the requested key,
while the restart node indicates where to restart from, if at all.
The difference is visible for the tree with a single node, which
crossed no backup node and which will return NULL in the second. That's
not an issue with next/prev since by definition the looked up node has
to exist, so if there's a single node there can't be a next nor a prev,
but for the le/ge etc variants it does make a huge difference since now
one may tell if a matching node was found and the other one may tell if
we can try our chance again on the other branch.

Overall the change is pretty minimal.