EXP: loop inside the function for next/prev
That's only for u64 for now. The result is slightly smaller but also slower.

The construct is not very convenient but allows to get rid of annoying
args. It also shows what we really need to keep (a pointer to the last
node, and a copy of the pxor, though that's one for each type). It also
requires an early check before entering the loop to make sure we detect
the leaf that depends from the same node, otherwise we can't detect it
later.

Maybe at some point we should instead define a context to be reused
between iterations instead of passing that many pointers. The only
issue is that it will no longer be possible to eliminate useless writes
to such variables we don't need.