EXP: descent: finally just return the return node
Now we only return the return node on KNX/KPR, and pass a pointer to
it as the root pointer for next lookup using NXT/PRV, which will do the
right thing. It will measure the distance between the node's two branches
again, but saves us from having to save the intermediary XOR values. It
still ends up being as fast as the original without requiring access to
the left/right branches.