MINOR: descent: add new CB_WM_GE/GT/LE/LT walk methods
This adds the new walk methods and tests the final node accordingly.
Indeed, while the descent must only follow the bits, the final check
used to decide whether we'll return the node or NULL must match the
requested method. Without this the caller would have to check the
node's key again, which would make no sense.

The way it's done is that we check the method then compare the keys
using the corresponding test (e.g. == for equality, >= for ge, etc).
For strings/binary blocks it's slightly more complicated because we
need to check the result of the comparison to know who's larger than
who. The tests were switched between left/right in a few places so
that the operator matches the method, because for no particular reason
the left-hand used to be the key and the right hand the node instead
of the opposite.