MINOR: descent: implement node address indexing
This indexing method consists in sorting the tree according to the node's
address instead of its contents. This was the very first implementation
of cbtree. It can be useful to look up entries known by their address when
a pointer serves as an identifier, as well as in memory allocators where
the node's address allows to defragment memory when releasing it; a
24-byte node can contain l,r,size on 64-bit systems and provide 16-byte
aligned blocks of 8-16 byte preceeded by their size. Using arenas and
32-bit sizes allows further shrinking, though that will require to use
different types.

The node's pointer is passed in key_ptr here, and internally all
operations are made on uintptr_t, with the xor stored in pxor64.