MINOR: private: make a separate function to compute the xor between two branches
At several places in the debugging code and in the dump code, we need
to compute the xor between two branches (or the common length). This
causes excessive code duplication and even limits the ease of implementing
dump functions for certain types.

This commit implements this missing function by extracting the code from
dbg() and generalizing it to all types and supporting any branch with a
key or both branches together.