MINOR: private: make the default dump function also take a root dump function
The root is a special case, it needs to draw an arrow to the first node
or leaf. It's painful to have to implement specific code for this in the
caller, not knowing the implementation details.

Now we're passing a root dump function, but this also requires that the
prototype is changed a little bit so that the root itself is passed,
instead of a pointer to the first node. But overall it's not bad, it's
even cleaner this way.