MINOR: scalar: use the default dump functions to dump the tree
Now we can get rid of the specific code and rely on the generic one
instead to dump the tree. Doing so also gets rid of older code that
once used to support duplicates as well as leaf detection using tagged
pointers, but since for now they've been dropped, there's no point in
keeping that code.