MINOR: private: pass kofs from all calling functions
Now the size of the struct node is passed by exported functions, which
means that it will now be possible to create new functions taking it as
an argument. The size of the code has not changed.

Note that in the callers, a number of functions still retrieve a key from
a node via a call to containerof() involving ceb_node_key!