MINOR: private: create macros to help with multi-mode function declarations
The new CEB_FDECL* macros will be used to create two variants of a same
function, one with sizeof(struct ceb_node) and one with "kofs" as the
key offset. The trick is to make both of them rely on the same static
inline function via a forward declaration, and to end with the prototype
of that static function whose body immediately follows. This way using a
single decorated function declaration, two functions are created. The
"root" argument is always the first one and the "kofs" argument, when
present, is just after the root.