OPTIM: descent: prefetch the 4 N+1 nodes
The nature of the tree which requires to compare next nodes' values
makes it very sensitive to prefetching.

A test with the LRU cache showed that 100M u64 lookups in a 64k entries
table takes 15.1s with eb64, 17.1s with cebu64, and 14.4s with extra
prefetches. That's a saving of 27ns per lookup for data that are already
in the L3 cache.

Even deletes are faster, as seen below with a 33% miss test:

  eb64:            15.429
  cebu64-base:     23.752
  cebu64+prefetch: 19.928

That's 38ns saved on average.

Finally, a larger test with 100M u64 lookups in a 2^20 entries table
with no misses gives:

  eb64:            64.6s     (115MB)
  cebu64-base:     60.8s     (82MB)
  cebu64+prefetch: 54.0s     (82MB)

=> that's 100ns saved per lookup compared to eb64.