REORG: extract the code for unique keys to cebutree-prv.h
At the moment some code will be duplicated as it's not certain that
the _cebu_descend() function will be adaptable for multiple keys. Let's
move it to a separate file with the cebu prefix since the .c already
have it anyway. This will leave some room for the more generic function
in cebtree-prv.h.