MEDIUM: extend the ceb tree dump API to pass a subtree number
For now it's not used, this change only updates the API without any
functional change, because there are multiple stacked layers, in order to
ease review and adaptations. The cebu32/64/a/is/l/s variants were updated
to take an extra parameter which contains the subtree number (callers may
just pass zero when unknown). This also impacts the tests, obviously.