MEDIUM: make the tree dump consider sub trees
It's always difficult during development to follow the steps of various
operations applied to a tree. Till now this had to be emitted to multiple
files which were then aggregated.

Here with this new approach, we're suffing node names in the graph with
a sub-graph number, allowing to create multiple independent subgraphs in
the same graph. This is very convenient to observe the steps. In order
to keep the compatibility with existing code, only an extra argument "sub"
was added to the function (the subtree number), and when zero, the whole
tree is dumped at once like before. When non-zero, only a subtree is dumped,
and the prologue/epilogue have to be emitted separately, by passing
root=NULL when calling the function (label defines the prologue, no label
the epilogue):

   sub  root label
    0   NULL NULL   epilogue only (closing brace and LF)
    0   NULL text   prologue with <text> as label
    0   tree *      prologue+tree+epilogue at once
   N>0  tree *      only the tree, after a prologue and before an epilogue

Now testceb32.c writes one tree per inserted number.