MINOR: private: start to consider the possibility of duplicates in dumps
Duplicates will be represented as lists. They will have as a nice
property that all of their neighbors will have the exact same key
and that their neighbors always exist. However there is a special
case for the 2nd entry, which is that it's the only one that may
have both left and right pointers pointing to the same other node,
which currently confuses the node dumping code which takes it for
the first element. Similarly the head of a link could be wrongly
set to "n" instead of "l" due to the heuristics used to detect them.

It's sufficient to add as a condition that the node must also be
itself for the first element, or to check for the branches values
when emitting the link type, in order to avoid this issue.

Note that there's a limitation: duplicate strings are not properly
reported due to the way they're compared (number of bits instead of
a real xor, i.e. the value is inverted). Some string dups are shown
as pure leaves. It only affects the dump though.