MEDIUM: extend the tree dump API again to pass a callback to show duplicates
ceb_default_dump_tree() was extended again. Now it can walk over
duplicates and print them. The walk method is quite simple, it's
only a walk left until a tree leaf is met (which is detected by
either the two pointers pointing to itself or having two different
neighbor keys).

A default dump function was added. It uses a beige color which is not
much different from the leaf one (yellow) and eases identification.
It tries hard to place origin and destination ports of the arrows
depending on the first/last nodes so that they never cross. The dups
are printed like a stair with curved arrows that are easy to spot. It
has been tested on experimental code and even managed to spot a bug in
the early implementation, so it's important to have it early in the
development cycle.

It is done in a way that remains compatible with unique trees, which
justifies to update the generic dump functions.