MEDIUM: descent: stop on duplicates during the descent
Till now we would stop on duplicates when the xor between the branches
increases again, but that's not compatible with the use of duplicates.
Indeed, this can bring us too far, which is particularly visible during
insertion where the new node is inserted at the bottom of the dup list.

Just like we're doing with the detection of leaves, we need to stop as
soon as we're certain we're on a dup. That's actually guaranteed by a
xor value of zero, or a full-size equality between the two sides. And
we always have that info since the xor between the two sides is always
calculated.

We also need to make the nodeless node check more accurate for the
keyless walk otherwise the next walk can wrongly stop on a two-dups
sub-tree believing it found the nodeless node. A performance test would
be useful here to see if checking (l==r && l==k) could be faster than
(l==k && r==k) due to the first condition being rare.

Since this adds extra tests, it might possibly slightly decrease the
performance in the descent, but given that it's bound by the memory
bandwidth anyway, it shouldn't have much impact. Also for the string
case we could later improve this by folding the two consecutive tests
by letting the compiler know that (ssize_t)plen is always positive, but
that's out of scope for now.