MEDIUM: insert: support inserting duplicates
When _ceb_descend() returns non-null, we know a node exists with the same
key at this address. By checking the is_dup return value, we can determine
if that node is a regular leaf or a duplicate entry (which for the rest of
the code acts as a nodeless leaf). This permits to insert a new dup, either
the first one (when there was a regular leaf) or as an extra dup entry
completing an existing one. Dumps show that it seems to be working correctly.