MINOR: descent: add a return pointer to report duplicates
If a duplicate is found, it will be returned in ret_is_dup now. This will
be useful later with some operations such as insert. We also take this
opportunity for skipping the duplicate test in the descent when not
needed. This means that callers that care about duplicates will need to
be careful to pass &is_dup there, but it's worth it in order to save
precious cycles with unique keys.