MINOR: tests: make testceb32 dump all entries at the end
The test now emits all entries in direct and reverse order using
ceb_first()+ceb_next() and ceb_last()+ceb_prev(). It's produced as
a comment so that it can remain in the dot file for later checks as
well.

Note that at the moment only unique keys are dumped (with the last
one of each dup subtree).