MINOR: private: add new _ceb_{next,prev} functions supporting duplicates
These new functions properly visit the dups when there are for the current
value. The check is not exactly the same for both directions since the dups
start with the last inserted one, but thanks to is_dup we can always figure
where we are and how to continue without having to compare key values.

It's worth noting that the descent code uses is_dup to figure whether or not
to check for duplicates, and that as such we have to pass it even if we don't
need the return value (first call of next, last call of prev).