MINOR: private: add new _ceb_{next,prev} functions supporting duplicates
These new functions properly visit the dups when there are for the current
value. The check is not exactly the same for both directions since the dups
start with the last inserted one, but thanks to is_dup we can always figure
where we are and how to continue without having to compare key values.