MINOR: private: create all dup-capable functions from the unique-only ones
This introduces _ceb_insert(), _ceb_first(), _ceb_last(), _ceb_next(),
_ceb_prev(), _ceb_delete(), _ceb_lookup(), _ceb_lookup_ge(),
_ceb_lookup_gt(), _ceb_lookup_le(), and _ceb_lookup_lt(), which are for
now perfect duplicates of the other ones.

The functions have been placed before the other ones because it's likely
that a number of the unique-only ones will eventually disappear as they
will be identical (or identical enough not to care about special cases).