BUG/MEDIUM: fix memory block first/last implementation's missing key length
The memory block must pass the key length with all operations, but this
argument was missing from the _first() and _last() operations because
the underlying functions did not expose it.

This commit extends _cebu_first() and _cebu_last() to take the key length
and pass it to key_u64 like other functions. It's just zero when not used.

The prototypes for cebub_* and cebuib_* were fixed to reflect this (actually
the _kofs variant was correct though the argument was ignored).

(cherry picked from commit 92b9b5c5f44c4ab0d2b3423da5fb223c0bb585a2)
Signed-off-by: Willy Tarreau <w@1wt.eu>