MEDIUM: tree-wide: rename all the cebu*.[ch] files to ceb*.[ch]
It now becomes quite clear that both the multi and the unique keys will
have very similar code and will be better handled in the same source
file than in multiple. For example there seems to be no reason to have
separate dump functions. Let's rename all exported files from cebu* to
ceb*. In order to ease porting, for now we're adding wrappers for include
files with the old name.