MINOR: cebuis: add missing dump function
The cebuis_default_dump() function was mentioned in the include file
but not implemented in the C file. Let's do it now. It's the same as
the cebus one except for the key type (and the digraph name).

(cherry picked from commit c6dbdca33db8082fe82b62b56b57e840d4fe1198)
Signed-off-by: Willy Tarreau <w@1wt.eu>