MINOR: private: fix the string dump descent
When dumping strings, since the xor_branches() value is directly taken
from string_equal_bits(), all the checks performed on it against zero to
detect an equality are reversed, because equal strings are returned as
-1. The same is true for equal_bits() by the way except that this time
the function returns the exact number of equal bits for a fixed key len.
There was already a difficulty caused by inverting all the xor checks in
the dump code. Better make xor_branches() return the complement of the
equal length so that the value continues to follow the same direction as
the distance between the values, and inter-xor comparisons as well as
checks against zero are always correct. It also allows to preserve the
split bit using ~xor for reporting on the nodes. This even removes some
code.