MEDIUM: private: factor the unique and dup versions of lookup_{ge,gt}()
The lookup_ge/gt() calls are very similar together, and also very similar
between dups and unique. The dup-aware one only has an extra check based
on is_dup. Passing its pointer as an extra argument to the _ceb_lookup_*()
function permits to condition the dup-specific code based on the pointer
passed by the caller. Thus we now rely on the caller to pass that value,
and we can get rid of _cebu_lookup_ge/gt(). It was verified that the
functions are exactly the same size before and after.