MEDIUM: private: factor the unique and dup versions of delete()
The delete() call only differs by the block dedicated to the special
handling of duplicates. Like with other functions, let's just let the
caller provide the is_dup variable that's used for this, and use that
element to distinguish between the two versions. It was verified that
the code size doesn't change with this modification, and _cebu_delete()
could be removed.